attempted refactor. train is running but perhaps with error as MCTS leaves are not being populated
