+++ lib/game/game.py
@@ -1,0 +1,8 @@
+"""Defining the interface for Game class
+"""
+
+from abc import ABC
+
+
+class BaseGame(ABC):
+    pass
