+++ lib/model.py
@@ -1,13 +1,9 @@
-import collections
 import numpy as np
 
 import torch
 import torch.nn as nn
 
-from lib import game, mcts
 
-
-OBS_SHAPE = (2, game.GAME_ROWS, game.GAME_COLS)
 NUM_FILTERS = 64
 
 
@@ -96,108 +92,3 @@ class Net(nn.Module):
         pol = self.conv_policy(v)
         pol = self.policy(pol.view(batch_size, -1))
         return pol, val
-
-
-def _encode_list_state(dest_np, state_list, who_move):
-    """
-    In-place encodes list state into the zero numpy array
-    :param dest_np: dest array, expected to be zero
-    :param state_list: state of the game in the list form
-    :param who_move: player index (game.PLAYER_WHITE or game.PLAYER_BLACK) who to move
-    """
-    assert dest_np.shape == OBS_SHAPE
-
-    for col_idx, col in enumerate(state_list):
-        for rev_row_idx, cell in enumerate(col):
-            row_idx = game.GAME_ROWS - rev_row_idx - 1
-            if cell == who_move:
-                dest_np[0, row_idx, col_idx] = 1.0
-            else:
-                dest_np[1, row_idx, col_idx] = 1.0
-
-
-def state_lists_to_batch(state_lists, who_moves_lists, device="cpu"):
-    """
-    Convert list of list states to batch for network
-    :param state_lists: list of 'list states'
-    :param who_moves_lists: list of player index who moves
-    :return Variable with observations
-    """
-    assert isinstance(state_lists, list)
-    batch_size = len(state_lists)
-    batch = np.zeros((batch_size,) + OBS_SHAPE, dtype=np.float32)
-    for idx, (state, who_move) in enumerate(zip(state_lists, who_moves_lists)):
-        _encode_list_state(batch[idx], state, who_move)
-    return torch.tensor(batch).to(device)
-
-
-def play_game(mcts_stores, replay_buffer, net1, net2,
-              steps_before_tau_0, mcts_searches, mcts_batch_size,
-              net1_plays_first=None, device="cpu"):
-    """
-    Play one single game, memorizing transitions into the replay buffer
-    :param mcts_stores: could be None or single MCTS or two MCTSes for individual net
-    :param replay_buffer: queue with (state, probs, values), if None, nothing is stored
-    :param net1: player1
-    :param net2: player2
-    :return: value for the game in respect to player1 (+1 if p1 won, -1 if lost, 0 if draw)
-    """
-    assert isinstance(replay_buffer, (collections.deque, type(None)))
-    assert isinstance(mcts_stores, (mcts.MCTS, type(None), list))
-    assert isinstance(net1, Net)
-    assert isinstance(net2, Net)
-    assert isinstance(steps_before_tau_0, int) and steps_before_tau_0 >= 0
-    assert isinstance(mcts_searches, int) and mcts_searches > 0
-    assert isinstance(mcts_batch_size, int) and mcts_batch_size > 0
-
-    if mcts_stores is None:
-        mcts_stores = [mcts.MCTS(), mcts.MCTS()]
-    elif isinstance(mcts_stores, mcts.MCTS):
-        mcts_stores = [mcts_stores, mcts_stores]
-
-    state = game.INITIAL_STATE
-    nets = [net1, net2]
-    if net1_plays_first is None:
-        cur_player = np.random.choice(2)
-    else:
-        cur_player = 0 if net1_plays_first else 1
-    step = 0
-    tau = 1 if steps_before_tau_0 > 0 else 0
-    game_history = []
-
-    result = None
-    net1_result = None
-
-    while result is None:
-        mcts_stores[cur_player].search_batch(
-            mcts_searches, mcts_batch_size, state,
-            cur_player, nets[cur_player], device=device)
-        probs, _ = mcts_stores[cur_player].get_policy_value(
-            state, tau=tau)
-        game_history.append((state, cur_player, probs))
-        action = np.random.choice(game.GAME_COLS, p=probs)
-        if action not in game.possible_moves(state):
-            print("Impossible action selected")
-        state, won = game.move(state, action, cur_player)
-        if won:
-            result = 1
-            net1_result = 1 if cur_player == 0 else -1
-            break
-        cur_player = 1-cur_player
-        # check the draw case
-        if len(game.possible_moves(state)) == 0:
-            result = 0
-            net1_result = 0
-            break
-        step += 1
-        if step >= steps_before_tau_0:
-            tau = 0
-
-    if replay_buffer is not None:
-        for state, cur_player, probs in reversed(game_history):
-            replay_buffer.append(
-                (state, cur_player, probs, result)
-            )
-            result = -result
-
-    return net1_result, step
